Move TrackPlacementArgs to CreateTrack.h
